RPT: r004 25
Getting simplecov to install using 'bundle install' took a while.
This is mainly because I was using windows as my host computer and
it requires 'make' which windows does not natively has. Therefore,
I had to create a Ubuntu VM in order to get simplecov to download and
install properly.