RPT: r007 30
Added a new test case that is a three element array and selects the middle element as the pivot point.
This test case lowered my mutant alive count from 41 to 39. Next going to try and make more test cases
to lower the alive count where possible.
At the start was not able to lower alive count more, but just kept trying different cases until it
occured.