RPT: r007 30
Created a test case that calls the quicksort function with a negative lo value and
returns a sorted array. This test case lowered the mutant alive count from 37 to 28.
Going to continue and attempt to lower the alive count even further.

At this point the test cases are trial and error because the mutant results can be
very ambiguous by changing constant values.