RPT: r007 22
Created a test case that calls quicksort with the same hi and lo value. This case
lowered the alive count from 28 to 24.
Ran into the issue of understanding the remaining warnings given by mutant. May
attempt to lower the alive count but believe that not many more can be killed.