RPT: r007 25
Created a test case that calls quicksort with an array if size 21, lo of 10 and hi of 20.
This test case lowered the alive count from 24 to 19.

There is not much documentation on mutant so attempting to come up with test cases to
satisfy the last couple warnings is a bit of a challenge.