RPT: r008 30
Still working on step 3. I was able to replicate the stdin in rspec as a user for one test.
Now that I am able to send in stdin data in rspec, I can proceed to attempt to check stdout
in a similar fashion.

I got stuck for a bit trying to send in stdin data to the program. There are many different ways that
are shown online and in the examples that I kept trying all of them and none of them seem to work.
Luckily the one example shown by Tommasa Pavese worked and fulfulled the sending of stdin
without modifying the nim.rb code.