RPT: r008 30
Working on steps 3 and 4 simultaneously. Step 3 in the sense that I still need to handle
stdout properly, and step 4 in the sense that I am making test to fulfull the 100%
coverage in simplecov. I am currently at 67.22% covered.

Ran into the issue of not handling stdout properly and test would just break because it would skip
sections. Working through trying to figure out how to handle stdout.