RPT: r008 23
Completed the stdout handling of step 3 and moving to step4 to achieve 100% coverage
in simplecov. Currently the simplecov covered is 61.47%.

Ran into the obstacle of understanding how rspec handles the expects for stdout. There
are many different ways to handled and similar to the stdin, all my previous attempts
failed. Luckily a post on reformatcode.com had a very simple solution that is also
displayed by Tommaso Pavese.