RPT: r008 20
Working on step 4. Improved the simplecov coverage to 79.74. In order to play an entire game,
I brute forced the inputs for the user. Need to see if there is any other way to go about
that.

Not any issues arose other than variable input depending on output has not been looked at.