RPT: r009 23
Downloaded and added gcd.rb in the r009 folder. Ran cucumber --init to setup the
appropriate file structure to run cucumber. Copied the features file into that directory.

When I attempted to run cucumber, it did not throw any errors or warnings even though
the instructions said it would. Need to look into why it is not throwing any issues.