RPT: r010 22
Downloaded prime_factors.rb, initialized the required files for cucumber and
copied over the prime.feature file to the correct location.

Ran into the issue of cucumber displaying a parser error. Turns out that if
the keyword 'Feature' begins with a lowercase 'f', cucumber is not able to
detect the tests.