RPT: r010 30
Created a prime_testing.rb that handles the step_definitions for the tests.
Also, modified the testing so that it is done by a table instead of the
original tedious way.

Need to handle the array return so that it is split into an int array
instead of a string.