RPT: r010 13
Completed the final step (step 3) of r010. This was converting the string into
an integer array so that we can compare the result of the function with what
we expect.

Ran into the issue of the first element resulting 0 while the rest were fine.
I solved this by stripping the quotations from the string and then converting
that into an integer array. It appeared that Ruby would escape the first quotation
which then when attempting to split the string, the quotation would try to convert to
an integer and thus resulting it to 0.