RPT: r011 27
Completed Step 1. Created the necessary files; Gemfile, beancounting.rb, bean_testing.rb,
acm00.feature and modified env.rb to link to the correct file. Also, in the testing
I created an input and output as specified in the question.

Have not considered how the input is going to be handled as of right now but I
believe the base work is done for input and output.