RPT: r011 30
Created beancounting.rb with initializer that takes in the string input of the user.
Also created/modified the step_definitions to pass in the input to an instance of
BeanCounting. Lastly, figuredout how to count the number of occurrances of a character
in a string. Will be using that to determine the winners.