RPT: r011 30
Modified the steps file to modify the input and remove the quatations of it.
Currently, I am figuring out how to properly split the guesses into the correct
Hash tables.

I had an issue of understanding and using the .split() function. When the input
was passed to the steps file, it treated it as an array of characters as opposed
to a string. Therefore, to properly split I had to use single quotes for characters.