RPT: r011 30
At the start I was stuck on understanding the problem at hand. Was confusing number of guesses with
total value guessed by the contestants. After reading it a couple of times, I was able to figure out
my error and calculations. Currently, creating a map to hold the number of beans in the jar and with
that I will be able to caluculate the shares.