RPT: r011 30
Calculated the results for each of the contestants and their shares. Constructed a string
that returned when message_count is called but the user and compares using cucumber.

When passing the output string to the steps_definition, newlines do not appear to be escaped
so I was stuck trying to figure out what was I doing wrong. I resulted in calling .split
on newlines and constructing a final string with the expected structure.